bot_posts/requirements.txt
python-telegram-bot==20.3
pydantic-settings==2.9.1
pydantic==2.11.5
httpx==0.24.0
dotenv==0.9.9
passlib[bcrypt]==1.7.4
